+++ integration-test/Cargo.toml
@@ -11,7 +11,7 @@ clap = { version = "3.1", features = ["derive"] }
 ckb-crypto = "0.105"
 ckb-hash = "0.105"
 ckb-jsonrpc-types = "0.105"
-ckb-sdk-otx = { git = "https://github.com/EthanYuan/ckb-sdk-rust.git", branch = "opentx_sign_tx_ethan", package = "ckb-sdk" }
+ckb-sdk = "=2.3.0"
 ckb-types = "0.105"
 env_logger = "0.9.3"
 hex = "0.4"
@@ -19,6 +19,7 @@ inventory = "0.2"
 lazy_static = "1.4.0"
 log = "0.4"
 once_cell = "1.10"
+rand = "0.8"
 secp256k1 = { version = "0.24", features = ["recovery"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
