+++ integration-test/src/help.rs
@@ -4,12 +4,12 @@ use super::const_definition::{
 };
 use crate::const_definition::CURRENT_OTX_POOL_SERVICE_PROCESS;
 use crate::utils::client::mercury_client::MercuryRpcClient;
+use crate::utils::instruction::command::run_command_spawn;
 use crate::utils::instruction::{ckb::generate_blocks, ckb::unlock_frozen_capacity_in_genesis};
 
 use config::{parse, AppConfig, ConfigFile};
 use otx_sdk::client::OtxPoolRpcClient;
 use utils::client::ckb_client::CkbRpcClient;
-use utils::instruction::command::run_command_spawn;
 
 use common::lazy::{
     ACP_CODE_HASH, CHEQUE_CODE_HASH, DAO_CODE_HASH, PW_LOCK_CODE_HASH, SECP256K1_CODE_HASH,
@@ -17,7 +17,7 @@ use common::lazy::{
 };
 
 use anyhow::Result;
-use ckb_sdk_otx::Address;
+use ckb_sdk::Address;
 use ckb_types::H256;
 
 use std::env;
