+++ integration-test/src/tests_otx_sighash_lock/blank_check.rs
@@ -1,5 +1,6 @@
 use crate::const_definition::{CKB_URI, MERCURY_URI, OTX_POOL_URI, SCRIPT_CONFIG};
 use crate::help::start_otx_pool;
+use crate::utils::client::ckb_cli_client::ckb_cli_transfer_ckb;
 use crate::utils::client::mercury_client::MercuryRpcClient;
 use crate::utils::instruction::ckb::aggregate_transactions_into_blocks;
 use crate::utils::instruction::ckb::dump_data;
@@ -15,12 +16,11 @@ use otx_sdk::address::build_otx_address_from_secp_address;
 use otx_sdk::build_tx::OtxBuilder;
 use otx_sdk::client::OtxPoolRpcClient;
 use otx_sdk::signer::{SighashMode, Signer};
-use utils::client::ckb_cli_client::ckb_cli_transfer_ckb;
 
 use core_rpc_types::{GetBalancePayload, JsonItem};
 
 use anyhow::{Ok, Result};
-use ckb_sdk_otx::Address;
+use ckb_sdk::Address;
 use ckb_types::prelude::Entity;
 use ckb_types::{
     bytes::Bytes,
