+++ integration-test/src/utils/client/ckb_cli_client.rs
@@ -1,7 +1,7 @@
-use crate::instruction::command::run_command_output;
+use crate::utils::instruction::command::run_command_output;
 
 use anyhow::Result;
-use ckb_sdk_otx::Address;
+use ckb_sdk::Address;
 use ckb_types::H256;
 
 use std::str::FromStr;
@@ -33,7 +33,7 @@ pub fn ckb_cli_transfer_ckb(address: &Address, capacity: usize) -> Result<H256> 
     H256::from_str(&tx_hash).map_err(Into::into)
 }
 
-pub fn ckb_cli_get_capacity(address: &Address) -> Result<f64> {
+pub fn _ckb_cli_get_capacity(address: &Address) -> Result<f64> {
     // ckb-cli
     // config --url http://127.0.0.1:8114"
     let (stdout, _) = run_command_output(
