+++ integration-test/src/utils/client/mod.rs
@@ -1,1 +1,2 @@
+pub mod ckb_cli_client;
 pub mod mercury_client;
