+++ integration-test/src/utils/instruction/mod.rs
@@ -1,2 +1,3 @@
 pub mod ckb;
+pub mod command;
 pub mod mercury;
