+++ otx-sdk/Cargo.toml
@@ -10,7 +10,7 @@ anyhow = "1.0"
 ckb-crypto = "0.105"
 ckb-hash = "0.105"
 ckb-jsonrpc-types = "0.105"
-ckb-sdk-otx = { git = "https://github.com/EthanYuan/ckb-sdk-rust.git", branch = "opentx_sign_tx_ethan", package = "ckb-sdk" }
+ckb-sdk = "=2.3.0"
 ckb-types = "0.105"
 jsonrpc-core = "18.0"
 reqwest = { version = "0.11", features = ["blocking", "json"] }
