+++ plugins-built-in/dust-collector/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 [dependencies]
 anyhow = "1.0"
 ckb-sdk = "=2.3.0"
-ckb-sdk-open-tx = { git = "https://github.com/EthanYuan/ckb-sdk-rust.git", branch = "opentx_sign_tx_ethan", package = "ckb-sdk"}
 ckb-stop-handler = "0.105"
 ckb-types = "0.105"
 dashmap = "5.4.0"
@@ -18,4 +17,5 @@ log = "0.4.17"
 config = { path = "../../util/config" }
 otx-format = { path = "../../otx-format" }
 otx-plugin-protocol = { path = "../../otx-plugin-protocol" }
+otx-sdk = { path = "../../otx-sdk" }
 utils = { path = "../../utils" }
