+++ utils/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod aggregator;
 pub mod build_tx;
 pub mod client;
-pub mod instruction;
-pub mod lock;
-pub mod wallet;
